Local bot test pics
I was having a hard time with my understanding as well as getting a hold of a decent IOT device. I have decided to switch to testing this with a local network made of virtual machines communicating through ssh and using default passwords to emulate a bot net. I will upload the actual code file and more testing as it is more refined.